Xcode 7b4 fixes
There is a change/bug that causes subclasses to no longer reference their superclass' exported name in the -Swift.h file. This causes a compile time issue. Work around is to no longer export the LogDesitnationBase with a different name. Also changed unit tests to use the properly exported ObjC names, as well as the THGLog macro